refactor: mv OrderingUtil to client
