+++ packages/client/package.json
@@ -71,11 +71,13 @@
         "@babel/preset-typescript": "^7.16.0",
         "@ethersproject/wordlists": "^5.5.0",
         "@types/express": "^4.17.13",
+        "@types/heap": "^0.2.28",
         "@types/lodash": "^4.14.175",
         "@types/node": "^16.10.6",
         "@types/node-fetch": "^2.5.12",
         "@types/split2": "^3.2.1",
         "@types/uuid": "^8.3.1",
+        "array-shuffle": "^2.0.0",
         "babel-loader": "^8.2.3",
         "babel-plugin-add-module-exports": "^1.0.4",
         "babel-plugin-lodash": "^3.3.4",
@@ -133,6 +135,7 @@
         "debug": "^4.3.2",
         "env-paths": "^2.2.1",
         "eventemitter3": "^4.0.7",
+        "heap": "^0.2.6",
         "idb-keyval": "^5.1.5",
         "lodash": "^4.17.21",
         "mem": "^8.1.1",
@@ -147,6 +150,7 @@
         "sqlite3": "^5.0.3",
         "streamr-client-protocol": "^12.0.0",
         "streamr-network": "^35.0.0",
+        "strict-event-emitter-types": "^2.0.0",
         "ts-toolbelt": "^9.6.0",
         "tsyringe": "^4.6.0",
         "uuid": "^8.3.2"
