+++ packages/client/test/unit/OrderedMsgChain.test.ts
@@ -1,13 +1,9 @@
 import assert from 'assert'
 
 import shuffle from 'array-shuffle'
-
-import OrderedMsgChain from '../../../src/utils/OrderedMsgChain'
-import StreamMessage from '../../../src/protocol/message_layer/StreamMessage'
-import GapFillFailedError from '../../../src/errors/GapFillFailedError'
-import MessageID from '../../../src/protocol/message_layer/MessageID'
-import MessageRef from '../../../src/protocol/message_layer/MessageRef'
-import { toStreamID } from '../../../src'
+import { MessageID, MessageRef, StreamMessage, toStreamID } from 'streamr-client-protocol'
+import OrderedMsgChain from '../../src/subscribe/ordering/OrderedMsgChain'
+import GapFillFailedError from '../../src/subscribe/ordering/GapFillFailedError'
 
 /**
  * Split an array into numChunks chunks.
@@ -366,6 +362,7 @@ describe('OrderedMsgChain', () => {
                     util.add(msg2)
                 }, 25)
             }
+
             if (to.timestamp === 4) {
                 util.add(msg4)
             }
@@ -388,6 +385,7 @@ describe('OrderedMsgChain', () => {
                         util.markMessageExplicitly(msg2)
                     }, 35)
                 }
+
                 if (to.timestamp === 4) {
                     setTimeout(() => {
                         util.markMessageExplicitly(msg4)
