+++ packages/protocol/package.json
@@ -26,21 +26,15 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@ethersproject/wallet": "^5.5.0",
-    "@types/debug": "^4.1.7",
-    "@types/heap": "^0.2.28",
     "@types/secp256k1": "^4.0.3",
     "@types/sinon": "^10.0.4",
-    "array-shuffle": "^2.0.0",
     "sinon": "^11.1.2",
     "streamr-test-utils": "^2.0.0",
     "typedoc": "^0.22.5"
   },
   "dependencies": {
     "@streamr/utils": "^0.0.1",
-    "debug": "^4.3.2",
-    "heap": "^0.2.6",
     "secp256k1": "^4.0.2",
-    "sha3": "^2.1.4",
-    "strict-event-emitter-types": "^2.0.0"
+    "sha3": "^2.1.4"
   }
 }
