+++ packages/protocol/src/errors/index.ts
@@ -1,13 +1,11 @@
 import InvalidJsonError from "./InvalidJsonError"
 import UnsupportedVersionError from "./UnsupportedVersionError"
-import GapFillFailedError from "./GapFillFailedError"
 import ValidationError from "./ValidationError"
 import StreamMessageError from "./StreamMessageError"
 
 export {
     InvalidJsonError,
     UnsupportedVersionError,
-    GapFillFailedError,
     ValidationError,
     StreamMessageError,
 }
