+++ packages/protocol/src/utils/index.ts
@@ -1,4 +1,3 @@
-import OrderingUtil from "./OrderingUtil"
 import StreamMessageValidator, { StreamMetadata } from "./StreamMessageValidator"
 import SigningUtil from "./SigningUtil"
 import { createTrackerRegistry, TrackerRegistry, SmartContractRecord } from "./TrackerRegistry"
@@ -8,7 +7,6 @@ import { KeyExchangeStreamIDUtils } from "./KeyExchangeStreamID"
 import { EthereumAddress, ENSName, ProxyDirection } from "./types"
 
 export {
-    OrderingUtil,
     StreamMessageValidator,
     StreamMetadata,
     SigningUtil,
